+++ docs/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [v0.2.0] - 2022-04-19
+
+### Added
+
+- backup all images from a Harbor server.
+- upload all images to another Harbor server.
+
+### Changed
+
+- Golang version to 1.18.
+
 ## [v0.1.0] - 2022-03-06
 
 ### Added
@@ -9,5 +20,6 @@
 - backup all images from a Nexus3 server.
 - upload all images to another Nexus3 server.
 
-[Unreleased]: https://github.com/030/p2iwd/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/030/p2iwd/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/030/p2iwd/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/030/p2iwd/releases/tag/v0.1.0